Create hiking cubs project
The idea behind this one, for possibly quarterly and its like to get one of the SDGs taken care of like the angel tree boxes its first to help with training directors, but also to help with possible surplus. What we worry is Surplus, that means if its not being sold, AND STILL GOOD, how can we help meet some of the needs. Even if that is an issue.